backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
deepface
pytesseract
numpy
pandas
opencv-python
pillow
groq
argon2-cffi
google-auth
requests
email-validator
tf-keras
tensorflow
